tests/wpt/web-platform-tests/client-hints/accept_ch_lifetime_same_origin_iframe.tentative.https.html

<html>
<title>Accept-CH-Lifetime test with same-origin iframe</title>
<meta name="timeout" content="long">
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<!--
Apart from this webpage, the test opens another html web page. One test is run
in this web page, and another in the second web page.
-->

<script>

// This test fetches resources/accept_ch_lifetime.html in a same origin iframe.
// The response headers to that webpage contain Accept-CH and Accept-CH-Lifetime
// headers.

// Fetching that webpage should cause the user-agent to persist origin
// preferences for the client hints specified in Accept-CH header for a
// duration specified in the Accept-CH-Lifetime header.

// Next, to verify if the origin preferences were persisted by the user
// agent, this test fetches resources/expect_client_hints_headers.html
// in a new window. Fetching of resources/expect_client_hints_headers.html
// verifies that the user agent actually sends the client hints in the request
// headers.

// Test is marked as tentative until https://github.com/whatwg/fetch/issues/726
// is resolved.

// First, verify the initial state to make sure that the browser does not have
// client hints preferences cached from a previous run of the test.
promise_test(t => {
  return fetch("echo_client_hints_received.py").then(r => {
    assert_equals(r.status, 200)
    // Verify that the browser did not include client hints in the request
    // headers when fetching echo_client_hints_received.py.
    assert_false(r.headers.has("device-memory-received"), "device-memory-received");
  });
}, "Precondition: Test that the browser does not have client hints preferences cached");

var acceptChLifetimeLoaded;

async_test(t => {
  acceptChLifetimeLoaded = t.step_func(() => {
    // Open a new window. Verify that the user agent attaches the client hints.
    var verify_win = window.open("resources/expect_client_hints_headers.html");
    assert_not_equals(verify_win, null, "Popup windows not allowed?");
  });

  window.addEventListener('message', t.step_func((e) => {
    if(!e.source.location.pathname.includes("expect_client_hints_headers.html")) {
      return;
    }
    if(typeof e.data != "string")
      return;
    assert_equals(e.data, "PASS");
    t.done();
  }));
}, "Loading of resources/expect_client_hints_headers.html did not finish.");
</script>

<!-- Fetching this webpage should cause user-agent to persist client hint
 preferences for the origin.-->
<iframe onload="acceptChLifetimeLoaded()" src="resources/accept_ch_lifetime.html"></iframe>

</body>
</html>
